[45] Redo DSA param tampering
The G=P+1 attack works when the other parameters are created normally
and the change to G is made afterwards. If verification is attempted
with the original P, Q, Y but with the altered G then the magic
signature will verify correctly.

On other words the key does not have to be generated with the malicious
G, but the verifier needs to be forced to use it.