[32] Timing attack statistics
Perform a large number of requests and plot a frequency distribution.
This commit includes example data and a plot.

(A full solution would automatically do this and compare e.g. the mode
of the distribution (or something more complex) to determine the full
hash, this is just a POC / demonstration.)