[42] Add option to use real ASN.1 processor
Adds an option to use Go's ASN.1 processor to create a "real" ASN.1
block. Enable with the -a option: 's' for simple that simply adds the
predetermined value (default) and 'p' to use the processor. The result
is identical whichever option is used.

Also increases the RSA key size used to 2048. Using a 1024 bit key
doesn't leave enough trailing zeros to calculate an appropriate perfect
cube value (at least with 8 0xff bytes, it might be possible using fewer
in the padding).